Remove Channel. Add Reset() and PrepareForAssert() methods
We need reset method as quite often test clients are created once for
test package and the same client is reused across multiple tests.
It might be handy to reset client so net effect of one test can be
evaluated.

In async usages of TritonD client we need to make sure we yield to
goroutines so that logging is actually complete
and that we synchronize assertion code so that race detector doesn't
complain about it.