Do not reuse ZMQ sockets that had failed to send the data
We see periodic spikes in ZMQ related errors (from fleet-api pings service).
The theory is that we get a ZMQ socket in a failed state and then reuse it,
thus getting a lot of failing sends in a row.
Close a failing socket as soon as we get the first error from it.